Code cleanup, Change type of function arguements and documenting the function calls

+++ module_eeprom/src/eeprom.h
@@ -27,25 +27,31 @@ interface i_eeprom_communication {
      * @brief Read the data from particular address in EEPROM.
      *
      * @param addr base address in EEPROM to read data
-     * @param no_of_bytes number of bytes to read sequentially from that base address
+     * @param no_of_bytes number of bytes to read from that base address
      * @param data buffer to fill with data
-     * @return data
      */
 
-    void    read(uint8_t addr, unsigned int no_of_bytes, uint8_t data[]);
+    void    read(uint8_t addr, uint8_t no_of_bytes, uint8_t data[]);
 
     /**
      * @brief Write the data using in EEPROM.
      *
      * @param addr address in EEPROM to write data from
      * @param data set of data to be written
-     * @param no_of_db number of data bytes to write from the set of data. (no_of_db = 1 for random write mode and no_of_db > 1 for page write mode)
+     * @param no_of_db number of data bytes to write from the set of data. (no_of_db = 1 for random write mode and 1 < no_of_db <= 8  for page write mode). The no_of_db should not be greater than 8.
      */
 
 
-    void    write(uint8_t addr, uint8_t data[no_of_db], size_t no_of_db);
+    void    write(uint8_t addr, uint8_t data[no_of_db], uint8_t no_of_db);
 };
 
+
+/**
+ * @brief Service to read and write data from 24LC01BT-I/LT EEPROM.
+ *
+ * @param i_eeprom eeprom communication interface for the eeprom_service.
+ * @param i2c_master_if communication interface with an I2C master component.
+ */
 void eeprom_service(server interface i_eeprom_communication i_eeprom, client interface i2c_master_if i2c);
 
 #endif /* EEPROM_H_ */
